chore: 데이터 삭제 api 수정
사용자 설정 카테고리 삭제 기능 추가
