+++ src/main/java/com/flow/flow/calendar/model/GetCalendarRes.java
@@ -1,4 +1,4 @@
-package com.flow.flow.src.calendar.model;
+package com.flow.flow.calendar.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,8 +7,8 @@ import lombok.Setter;
 @Getter
 @Setter
 @AllArgsConstructor
-public class GetTotalAmount {
-    private int total;
+public class GetCalendarRes {
+    private int date;
     private int isExp;
+    private int amount;
 }
-
