+++ src/main/java/com/flow/flow/setting/SettingProvider.java
@@ -1,9 +1,9 @@
 
-package com.flow.flow.src.setting;
+package com.flow.flow.setting;
 
 import com.flow.flow.config.BaseException;
-import com.flow.flow.src.setting.model.Form;
-import com.flow.flow.src.setting.model.GetFormRes;
+import com.flow.flow.setting.model.Form;
+import com.flow.flow.setting.model.GetFormRes;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
