+++ src/main/java/com/flow/flow/setting/SettingService.java
@@ -1,13 +1,11 @@
-package com.flow.flow.src.setting;
+package com.flow.flow.setting;
 
 import com.flow.flow.config.BaseException;
-import com.flow.flow.src.setting.model.PatchBudgetReq;
+import com.flow.flow.setting.model.PatchBudgetReq;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 import static com.flow.flow.config.BaseResponseStatus.*;
